ORS_Agent_MCP/pyproject.toml
[project]
name = "model-context-protocol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "fastapi-mcp>=0.3.4",
    "fastmcp>=2.7.0",
    "langchain-groq>=0.3.2",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-mistralai>=0.2.10",
    "langchain-openai>=0.3.21",
    "langgraph>=0.4.7",
    "langgraph-checkpoint-sqlite>=2.0.10",
    "loguru>=0.7.3",
    "openrouteservice>=2.3.3",
    "pip>=25.1.1",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]
